requirements.txt
anyio==3.6.1
Babel==2.10.3
click==8.1.3
dnspython==2.2.1
graphql-core==3.2.1
h11==0.13.0
idna==3.3
iso8601==1.0.2
Pygments==2.13.0
pymongo==4.2.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytimeparse==1.1.8
pytz==2022.2.1
six==1.16.0
sniffio==1.2.0
starlette==0.20.4
strawberry-graphql==0.126.2
typing-extensions==4.3.0
uvicorn==0.18.3
zulu==2.0.0
